fix(pitboss-core): FailureReason layering, libgit2 cache, SessionStore::open (#188)
Three medium-severity follow-ups to the earlier #188 audit:

* **FailureReason layering (#188 M1)** — moved `classify`, all `match_*`
  helpers, `parse_reset_timestamp` / `parse_12h_time` /
  `month_from_abbrev`, and `excerpt` from `pitboss-cli::dispatch::
  failure_detection` to a new `pitboss_core::failure_classify` module.
  `FailureReason` already lived here; the constructor now lives next
  to it. Downstream consumers (the new `pitboss-web` console, future
  external integrations) can classify a log blob without depending
  on the CLI crate. `ApiHealth`, `SpawnGateReason`,
  `detect_failure_reason` (file IO), and `broadcast_worker_failed`
  (CLI control protocol) stay in CLI. All 13 classify-side tests
  moved; the CLI module keeps its `read_tail` / `ApiHealth` /
  `broadcast_worker_failed` tests.

* **libgit2 worktree-branch cache (#188 M2)** — `prepare`'s
  branch-conflict pre-check used to call `Repository::open(wt.path())`
  per existing worktree, dominating prepare time on operators with
  heavy worktree usage. Each worktree's HEAD is a plain text file at
  `<repo>/.git/worktrees/<name>/HEAD` containing
  `ref: refs/heads/<branch>`; the new `head_branch_for_worktree`
  reads it directly with no libgit2 open. Fallback path
  `head_branch_via_open` still handles detached-HEAD and unusual
  layouts. Three regression tests cover the cache reader.

* **`SessionStore::open` (#188 M3)** — added a uniform `fn open(path:
  &Path) -> Result<Box<dyn SessionStore>, StoreError>` to the trait.
  `JsonFileStore::new` was infallible and `SqliteStore::new` was
  fallible, so a polymorphic constructor was previously impossible —
  call sites that wanted to hand a store to something parameterized
  over `Arc<dyn SessionStore>` had to manually dispatch on the
  backend kind. The trait method returns `Box<dyn SessionStore>`
  because Rust trait-object methods can't return `Self`; callers
  `Arc::from(store)` it directly.

Refs #188.
