+++ crates/pitboss-core/src/worktree/manager.rs
@@ -65,13 +65,18 @@ impl WorktreeManager {
         }
 
         if let Some(bname) = &branch_name {
+            // #188 M2: avoid `Repository::open(wt.path())` per worktree —
+            // each open does a libgit2 discover() pass and dominates
+            // `prepare` time on operators with many active worktrees.
+            // Each worktree's HEAD is a plain text file at
+            // `<repo>/.git/worktrees/<name>/HEAD` containing
+            // `ref: refs/heads/<branch>`; read it directly. Falls back to
+            // the Repository::open path only on read/parse failure so
+            // detached-HEAD or symbolic-ref forms still classify.
             for wt_name in repo.worktrees()?.iter().flatten() {
                 let wt = repo.find_worktree(wt_name)?;
-                let wt_repo = Repository::open(wt.path())?;
-                let checked_out = wt_repo
-                    .head()
-                    .ok()
-                    .and_then(|h| h.shorthand().map(str::to_string));
+                let checked_out = head_branch_for_worktree(repo_root, wt_name)
+                    .or_else(|| head_branch_via_open(wt.path()));
                 if checked_out.as_deref() == Some(bname.as_str()) {
                     return Err(WorktreeError::BranchConflict {
                         branch: bname.clone(),
@@ -146,4 +151,79 @@ fn sibling_path(repo_root: &Path, name: &str) -> PathBuf {
         .file_name()
         .map_or_else(|| "repo".to_string(), |s| s.to_string_lossy().to_string());
     parent.join(format!("{base}-{name}"))
+}
+
+/// Cheap path-cache lookup for a worktree's currently checked-out branch.
+///
+/// Reads the plain-text HEAD file at
+/// `<repo_root>/.git/worktrees/<wt_name>/HEAD` directly instead of opening
+/// a fresh libgit2 `Repository` per worktree (which dominates
+/// `prepare` time on operators with many active worktrees — #188 M2).
+///
+/// Returns the branch shorthand on `ref: refs/heads/<branch>` form;
+/// `None` for detached HEAD, missing files, or any other shape — the
+/// caller's existing `Repository::open` path picks those up.
+fn head_branch_for_worktree(repo_root: &Path, wt_name: &str) -> Option<String> {
+    let head_file = repo_root
+        .join(".git")
+        .join("worktrees")
+        .join(wt_name)
+        .join("HEAD");
+    let contents = std::fs::read_to_string(&head_file).ok()?;
+    let line = contents.lines().next()?;
+    let suffix = line.strip_prefix("ref: refs/heads/")?;
+    Some(suffix.trim().to_string())
+}
+
+/// Fallback for the cache miss path: the older `Repository::open`-based
+/// branch-lookup. Kept so the cheap-cache path can degrade gracefully
+/// when the HEAD file is in an unexpected form (symref, alternate
+/// repository layout, etc.) without changing `prepare`'s behavior.
+fn head_branch_via_open(wt_path: &Path) -> Option<String> {
+    let wt_repo = Repository::open(wt_path).ok()?;
+    wt_repo
+        .head()
+        .ok()
+        .and_then(|h| h.shorthand().map(str::to_string))
+}
+
+#[cfg(test)]
+mod head_cache_tests {
+    use super::*;
+    use std::fs;
+    use tempfile::TempDir;
+
+    /// #188 M2 regression: the cheap-cache reader extracts the branch
+    /// shorthand from a plain `ref: refs/heads/<branch>` HEAD file
+    /// without going through libgit2.
+    #[test]
+    fn head_branch_for_worktree_reads_plain_text_head() {
+        let tmp = TempDir::new().unwrap();
+        let admin = tmp.path().join(".git").join("worktrees").join("wt-x");
+        fs::create_dir_all(&admin).unwrap();
+        fs::write(admin.join("HEAD"), "ref: refs/heads/feat-cool\n").unwrap();
+        let got = head_branch_for_worktree(tmp.path(), "wt-x");
+        assert_eq!(got.as_deref(), Some("feat-cool"));
+    }
+
+    /// Detached HEAD form should return None so the caller's fallback
+    /// path picks it up via libgit2.
+    #[test]
+    fn head_branch_for_worktree_returns_none_for_detached_head() {
+        let tmp = TempDir::new().unwrap();
+        let admin = tmp.path().join(".git").join("worktrees").join("wt-y");
+        fs::create_dir_all(&admin).unwrap();
+        fs::write(admin.join("HEAD"), "deadbeefdeadbeefdeadbeefdeadbeef\n").unwrap();
+        let got = head_branch_for_worktree(tmp.path(), "wt-y");
+        assert!(got.is_none(), "detached HEAD should not match the prefix");
+    }
+
+    /// Missing HEAD file (worktree pruned mid-walk) returns None,
+    /// not a panic.
+    #[test]
+    fn head_branch_for_worktree_returns_none_when_missing() {
+        let tmp = TempDir::new().unwrap();
+        let got = head_branch_for_worktree(tmp.path(), "no-such-worktree");
+        assert!(got.is_none());
+    }
 }
